Add files via upload
The 2For1 strategy is quite famous in basketball. A head rolling kind of stuff for planning for coaches at the end of each quarters so as to have differential gains. 
Basketball each possesions is of 24 seconds. So major plannings of teams tend to use each possesions so as to minimize chances of opponent getting possesions.
  Majorly teams plan for switching possessions in last 40 seconds so that the team my get possession for more number of times as most people think 2 chance for scoring is always better than 1. I collected data of more than 6000 quarters of NBA matches. 
Main point is to look at differential gain of scores between sides in last 40 seconds of quarters. I analysed also performances of individual players towards the end so that opponents may put a look whether it is advantageous to loose possesions for 2for1 gain or not. 
I anlysed using haet map matrix feature of seaborn to study clck-by-clock analysts of average gains in points.